feat: multi-layer posture detection
Combine hog (dlib), mtcnn and self-trained model.
Hog has the fastest speed but needs front faces, so it's used
as the first layer on clear images; then when the angle is large
enough that hog fails, mtcnn is used. It's so slow that we can't
have it as the first layer; last, when the above 2 detections really
can get a face, we use the model.